+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.2.20] - 2024-06-03
+
+### Changed
+- [matter.js]: Update to @project-chip/matter-node.js v. 0.9.1
+- [matterbridge]: Updated dependencies
+
+### Fixed
+- [matterbridge]: Log level of Plugin already configured is now info
+
+<a href="https://www.buymeacoffee.com/luligugithub">
+  <img src="./yellow-button.png" alt="Buy me a coffee" width="120">
+</a>
+
 ## [1.2.19] - 2024-06-01
 
 ### Breaking change on Matterbridge start!